+++ tests/tt_metal/tt_metal/dispatch/random_program_fixture.hpp
@@ -14,7 +14,7 @@
 #include <tt-metalium/circular_buffer_constants.h>
 #include <tt-metalium/kernel_types.hpp>
 #include <tt-metalium/tt_backend_api_types.hpp>
-#include <tt-metalium/semaphore.hpp>
+#include "impl/buffers/semaphore.hpp"
 #include "dispatch_test_utils.hpp"
 
 namespace tt::tt_metal {
