+++ tt_metal/CMakeLists.txt
@@ -91,7 +91,6 @@ set(TT_METAL_PUBLIC_API
     api/tt-metalium/program_descriptors.hpp
     api/tt-metalium/routing_table_generator.hpp
     api/tt-metalium/runtime_args_data.hpp
-    api/tt-metalium/semaphore.hpp
     api/tt-metalium/shape.hpp
     api/tt-metalium/shape2d.hpp
     api/tt-metalium/shape_base.hpp
